@@ -10,26 +10,16 @@
     {
       "name": "Jozsef Geng",
       "email": "gengjozsef86@gmail.com"
-    },
-    {
-      "name": "Markus Nörenberg",
-      "email": "noerenberg.markus@gmail.com"
     }
   ],
-  "scripts": {
-    "ci": "make ci",
-    "codeception": "make codeception",
-    "phpcs": "make phpcs",
-    "phpstan": "make phpstan"
-  },
   "require": {
-    "php": ">=7.3",
-    "spryker/api": "^0.1.0 || ^0.2.0 || ^0.3.0",
+    "php": ">=8.0",
+    "spryker/api": "^0.4.0",
     "spryker/api-query-builder": "^0.1.0",
     "spryker/sales": "^8.6.0 || ^10.0.0 || ^11.0.0"
   },
   "require-dev": {
-    "fond-of-codeception/spryker": "^1.0",
+    "fond-of-codeception/spryker": "^1.0 || ^2.0.0",
     "spryker/code-sniffer": "*",
     "spryker-sdk/phpstan-spryker": "*"
   },
