@@ -2,21 +2,18 @@
 
 namespace FondOfSpryker\Zed\OrderApi\Business;
 
-use FondOfSpryker\Zed\OrderApi\Business\Mapper\TransferMapper;
-use FondOfSpryker\Zed\OrderApi\Business\Mapper\TransferMapperInterface;
 use FondOfSpryker\Zed\OrderApi\Business\Model\OrderApi;
 use FondOfSpryker\Zed\OrderApi\Business\Model\OrderApiInterface;
 use FondOfSpryker\Zed\OrderApi\Business\Model\Validator\OrderApiValidator;
 use FondOfSpryker\Zed\OrderApi\Business\Model\Validator\OrderApiValidatorInterface;
+use FondOfSpryker\Zed\OrderApi\Dependency\Facade\OrderApiToApiFacadeInterface;
 use FondOfSpryker\Zed\OrderApi\Dependency\Facade\OrderApiToSalesFacadeInterface;
-use FondOfSpryker\Zed\OrderApi\Dependency\QueryContainer\OrderApiToApiQueryBuilderQueryContainerInterface;
-use FondOfSpryker\Zed\OrderApi\Dependency\QueryContainer\OrderApiToApiQueryContainerInterface;
 use FondOfSpryker\Zed\OrderApi\OrderApiDependencyProvider;
 use Spryker\Zed\Kernel\Business\AbstractBusinessFactory;
 
 /**
  * @method \FondOfSpryker\Zed\OrderApi\OrderApiConfig getConfig()
- * @method \FondOfSpryker\Zed\OrderApi\Persistence\OrderApiQueryContainer getQueryContainer()
+ * @method \FondOfSpryker\Zed\OrderApi\Persistence\OrderApiRepository getRepository()
  */
 class OrderApiBusinessFactory extends AbstractBusinessFactory
 {
@@ -27,19 +24,9 @@ class OrderApiBusinessFactory extends AbstractBusinessFactory
     {
         return new OrderApi(
             $this->getSalesFacade(),
-            $this->getApiQueryContainer(),
-            $this->getApiQueryBuilderQueryContainer(),
-            $this->getQueryContainer(),
-            $this->createTransferMapper(),
+            $this->getApiFacade(),
+            $this->getRepository(),
         );
-    }
-
-    /**
-     * @return \FondOfSpryker\Zed\OrderApi\Business\Mapper\TransferMapperInterface
-     */
-    public function createTransferMapper(): TransferMapperInterface
-    {
-        return new TransferMapper();
     }
 
     /**
@@ -51,19 +38,11 @@ class OrderApiBusinessFactory extends AbstractBusinessFactory
     }
 
     /**
-     * @return \FondOfSpryker\Zed\OrderApi\Dependency\QueryContainer\OrderApiToApiQueryContainerInterface
+     * @return \FondOfSpryker\Zed\OrderApi\Dependency\Facade\OrderApiToApiFacadeInterface
      */
-    protected function getApiQueryContainer(): OrderApiToApiQueryContainerInterface
+    protected function getApiFacade(): OrderApiToApiFacadeInterface
     {
-        return $this->getProvidedDependency(OrderApiDependencyProvider::QUERY_CONTAINER_API);
-    }
-
-    /**
-     * @return \FondOfSpryker\Zed\OrderApi\Dependency\QueryContainer\OrderApiToApiQueryBuilderQueryContainerInterface
-     */
-    protected function getApiQueryBuilderQueryContainer(): OrderApiToApiQueryBuilderQueryContainerInterface
-    {
-        return $this->getProvidedDependency(OrderApiDependencyProvider::QUERY_CONTAINER_API_QUERY_BUILDER);
+        return $this->getProvidedDependency(OrderApiDependencyProvider::FACADE_API);
     }
 
     /**
