@@ -3,7 +3,6 @@
 namespace FondOfSpryker\Zed\OrderApi\Business;
 
 use Generated\Shared\Transfer\ApiCollectionTransfer;
-use Generated\Shared\Transfer\ApiDataTransfer;
 use Generated\Shared\Transfer\ApiItemTransfer;
 use Generated\Shared\Transfer\ApiRequestTransfer;
 use Spryker\Zed\Kernel\Business\AbstractFacade;
@@ -14,19 +13,15 @@ use Spryker\Zed\Kernel\Business\AbstractFacade;
 class OrderApiFacade extends AbstractFacade implements OrderApiFacadeInterface
 {
     /**
-     * {@inheritDoc}
-     *
-     * @api
-     *
-     * @param \Generated\Shared\Transfer\ApiDataTransfer $apiDataTransfer
+     * @param \Generated\Shared\Transfer\ApiRequestTransfer $apiRequestTransfer
      *
      * @return array
      */
-    public function validate(ApiDataTransfer $apiDataTransfer): array
+    public function validate(ApiRequestTransfer $apiRequestTransfer): array
     {
         return $this->getFactory()
             ->createOrderApiValidator()
-            ->validate($apiDataTransfer);
+            ->validate($apiRequestTransfer);
     }
 
     /**
